# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WeaponSummary {
  slug: string;
  title: string;
  family: string;
  manufacturer: string;
  country: string;
  ammunition: string;
  kind: string;
  summary: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  yearFrom?: number | null;
  /** @nullable */
  yearTo?: number | null;
  /** @nullable */
  variantOf?: string | null;
  variantCount: number;
}

export interface MediaImage {
  url: string;
  alt: string;
  caption: string;
  /** @nullable */
  sourceUrl?: string | null;
  sourceName: string;
}

export type ResourceLinkKind = typeof ResourceLinkKind[keyof typeof ResourceLinkKind];


export const ResourceLinkKind = {
  Видео: 'Видео',
  Статья: 'Статья',
  Музей: 'Музей',
  Архив: 'Архив',
  Схема: 'Схема',
} as const;

export interface ResourceLink {
  title: string;
  url: string;
  kind: ResourceLinkKind;
  /** @nullable */
  description?: string | null;
  sourceName: string;
}

export type WeaponDetail = WeaponSummary & {
  description: string;
  principle: string;
  safetyNote: string;
  variants: WeaponSummary[];
  images: MediaImage[];
  resources: ResourceLink[];
};

export interface Comment {
  id: number;
  weaponSlug: string;
  nickname: string;
  body: string;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 2
     * @maxLength 32
     */
  nickname: string;
  /**
     * @minLength 3
     * @maxLength 1000
     */
  body: string;
}

export interface WeaponMeta {
  families: string[];
  ammunition: string[];
  countries: string[];
  manufacturers: string[];
}

export interface ErrorResponse {
  error: string;
}

export type SearchQueryParameter = string;

export type FamilyQueryParameter = string;

export type AmmunitionQueryParameter = string;

export type CountryQueryParameter = string;

export type ManufacturerQueryParameter = string;

export type ListWeaponsParams = {
/**
 * @maxLength 120
 */
q?: SearchQueryParameter;
/**
 * @maxLength 80
 */
family?: FamilyQueryParameter;
/**
 * @maxLength 80
 */
ammunition?: AmmunitionQueryParameter;
/**
 * @maxLength 80
 */
country?: CountryQueryParameter;
/**
 * @maxLength 120
 */
manufacturer?: ManufacturerQueryParameter;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};