# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  weaponSlug: string;
  nickname: string;
  body: string;
  createdAt: Date;
}