# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /**
     * @minLength 2
     * @maxLength 32
     */
  nickname: string;
  /**
     * @minLength 3
     * @maxLength 1000
     */
  body: string;
}