# lib/api-zod/src/generated/types/mediaImage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */

export interface MediaImage {
  url: string;
  alt: string;
  caption: string;
  /** @nullable */
  sourceUrl?: string | null;
  sourceName: string;
}