# lib/api-zod/src/generated/types/resourceLink.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceLinkKind } from './resourceLinkKind';

export interface ResourceLink {
  title: string;
  url: string;
  kind: ResourceLinkKind;
  /** @nullable */
  description?: string | null;
  sourceName: string;
}