# lib/api-zod/src/generated/types/weaponDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */
import type { MediaImage } from './mediaImage';
import type { ResourceLink } from './resourceLink';
import type { WeaponSummary } from './weaponSummary';

export type WeaponDetail = WeaponSummary & {
  description: string;
  principle: string;
  safetyNote: string;
  variants: WeaponSummary[];
  images: MediaImage[];
  resources: ResourceLink[];
};