# lib/api-zod/src/generated/types/weaponMeta.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */

export interface WeaponMeta {
  families: string[];
  ammunition: string[];
  countries: string[];
  manufacturers: string[];
}