# lib/api-zod/src/generated/types/weaponSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Арсенал — каталог справочных материалов об огнестрельном оружии
 * OpenAPI spec version: 0.1.0
 */

export interface WeaponSummary {
  slug: string;
  title: string;
  family: string;
  manufacturer: string;
  country: string;
  ammunition: string;
  kind: string;
  summary: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  yearFrom?: number | null;
  /** @nullable */
  yearTo?: number | null;
  /** @nullable */
  variantOf?: string | null;
  variantCount: number;
}